// backend/src/bhxh/enums/qd3176.enum.ts
export enum QD3176_XML_TYPE {
  XML1 = 'XML1',
  XML2 = 'XML2',
  XML3 = 'XML3',
  XML4 = 'XML4',
  XML5 = 'XML5',
  XML6 = 'XML6',
  XML7 = 'XML7',
  XML8 = 'XML8',
  XML9 = 'XML9',
  XML10 = 'XML10',
  XML11 = 'XML11',
  XML12 = 'XML12',
  XML13 = 'XML13',
  XML14 = 'XML14',
  XML15 = 'XML15',
}

export enum QD3176_XML_IMPORT_QUEUE {
  XML_IMPORT = 'qd3176-xml-import',
}

export enum QD3176_XML_IMPORT_JOB {
  PARSE_AND_CREATE = 'parse-and-create',
  SESSION_FINALIZE = 'session-finalize',
}